% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eA1c.R
\name{eA1c}
\alias{eA1c}
\title{Compute Estimated A1c}
\usage{
eA1c(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A numeric value representing eA1c
}
\description{
This function computes the estimated A1c, according to
the American Diabetes Association calculator
}
