% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMI.R
\name{GMI}
\alias{GMI}
\title{Compute Glycemic Management Indicator}
\usage{
GMI(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A numeric value representing GMI
}
\description{
This function computes the estimated GMI
}
