% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBGI_HBGI.R
\name{HBGI}
\alias{HBGI}
\title{Compute High Blood Glucose Index}
\usage{
HBGI(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A numeric value representing HBGI
}
\description{
This function computes the high blood glucose index
}
