% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGBI_HGBI.R
\name{HGBI}
\alias{HGBI}
\title{Compute High Blood Glucose Index}
\usage{
HGBI(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A numeric value representing HGBI
}
\description{
This function computes the high blood glucose index
}
