% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interday.R
\name{interdaycv}
\alias{interdaycv}
\title{Compute Interday Coefficient of Variation}
\usage{
interdaycv(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A numeric value representing interday cv
}
\description{
This function computes the interday coefficient of variation
}
