% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interday.R
\name{interdaysd}
\alias{interdaysd}
\title{Compute Interday Standard Deviation}
\usage{
interdaysd(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A numeric value representing interday sd
}
\description{
This function computes the interday standard deviation
}
