% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday.R
\name{intradaycv}
\alias{intradaycv}
\title{Compute Intraday Coefficient of Variation}
\usage{
intradaycv(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A data frame containing the mean, median, and standard deviation
of the intraday coefficients of variation.
}
\description{
This function computes the intraday coefficient of variation
summary statistics: mean, median, standard deviation of all days in data
}
