% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday.R
\name{intradaysd}
\alias{intradaysd}
\title{Compute Intraday Standard Deviation}
\usage{
intradaysd(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A data frame containing the mean, median, and standard deviation
of the intraday standard deviations.
}
\description{
This function computes the intraday standard deviation summary statistics:
mean, median, standard deviation of all days in data
}
