% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_index.R
\name{J_index}
\alias{J_index}
\title{Compute J-index}
\usage{
J_index(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A numeric value representing J-index
}
\description{
This function computes J-index, a glycemic variability metrix
}
