% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBGI_HBGI.R
\name{LBGI_HBGI}
\alias{LBGI_HBGI}
\title{Compute Low Blood Glucose Index}
\usage{
LBGI_HBGI(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A data frame containing both the LBGI and HBGI values
}
\description{
This function computes the low blood glucose index
}
