% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGBI_HGBI.R
\name{LGBI}
\alias{LGBI}
\title{Compute Low Blood Glucose Index}
\usage{
LGBI(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A numeric value representing LGBI
}
\description{
This function computes the low blood glucose index
}
