% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGE.R
\name{MGE}
\alias{MGE}
\title{Compute Mean of Glycemic Excursions}
\usage{
MGE(df, sd = 1)
}
\arguments{
\item{df}{Data frame read through readfile}

\item{sd}{Standard deviation indicating glycemic excursion, default = 1}
}
\value{
A numeric value representing MAGE
}
\description{
This function computes the mean of glycemic excursions,
glycemic excursions indicated by standard deviation, default = 1
}
