% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glucose.R
\name{plot_glucose}
\alias{plot_glucose}
\title{Plot Glucose Data}
\usage{
plot_glucose(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\description{
This function plots glycemic excursions over the
time period in which data was collected.
}
