% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfile.R
\name{readfile}
\alias{readfile}
\title{Read in Data Frame}
\usage{
readfile(filename)
}
\arguments{
\item{filename}{.csv file of data frame to be read}
}
\value{
transformed data frame for further analysis
}
\description{
This function reads in a .csv with variable names
Timestamp..YYYY.MM.DDThh.mm.ss and Glucose.Value..mg.dL
}
