% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBGI_HBGI.R
\name{LBGI}
\alias{LBGI}
\title{Compute Low Blood Glucose Index}
\usage{
LBGI(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A numeric value representing LBGI
}
\description{
This function computes the low blood glucose index
}
\examples{
mydatafile <- system.file("extdata", "my_data_file.csv", package = "cgmquantify")
mydata <- readfile(mydatafile)
LBGI(mydata)
}
