% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGE.R
\name{MGN}
\alias{MGN}
\title{Compute Mean of Normal Glucose}
\usage{
MGN(df)
}
\arguments{
\item{df}{Data frame read through readfile}
}
\value{
A numeric value representing MGN
}
\description{
This function computes the mean of normal glucose,
glycemic excursions indicated by standard deviation, default = 1
}
